Migrate /tests to metatests
PR-URL: https://github.com/metarhia/metasync/pull/382
