+++ lib/collector.js
@@ -102,8 +102,8 @@ Collector.prototype.cancel = function(err) {
 };
 
 Collector.prototype.then = function(fulfill, reject) {
-  if (!fulfill) fulfill = common.emptyness;
-  if (!reject) reject = common.emptyness;
+  if (!fulfill) fulfill = common.emptiness;
+  if (!reject) reject = common.emptiness;
   this.onDone = (err, result) => {
     this.onDone = common.emptiness;
     if (err) reject(err);
