--- tests/all.js
@@ -1,14 +1,0 @@
-'use strict';
-
-const tests = [
-  'composition',
-  'composition.pause',
-  'composition.parallel',
-  'composition.sequential',
-  'composition.cancel',
-  'collector',
-];
-
-tests
-  .map(file => './' + file + '.js')
-  .forEach(require);
